+++ src/Handlers/PlayerSpawnedInWorld.cs
@@ -22,11 +22,13 @@ namespace Amnesia.Handlers {
                 switch (respawnType) {
                     case RespawnType.EnterMultiplayer: // first-time login for new player
                         _ = PlayerHelper.AddPositiveOutlookTime(player, Config.PositiveOutlookTimeOnFirstJoin);
+                        RefundHardenedMemory(clientInfo, player);
                         HandleStandardRespawnSteps(player);
                         break;
                     case RespawnType.JoinMultiplayer: // existing player rejoining
                         // grace period should continue only so long as you don't disconnect
                         player.Buffs.RemoveBuff(Values.BuffPostBloodmoonLifeProtection);
+                        RefundHardenedMemory(clientInfo, player);
                         HandleStandardRespawnSteps(player);
                         break;
                     case RespawnType.Died: // existing player returned from death
@@ -38,7 +40,7 @@ namespace Amnesia.Handlers {
                 log.Error("Failed to handle PlayerSpawnedInWorld event.", e);
             }
         }
-        
+
         /// <summary>
         /// Process steps common to enter/join/death.
         /// </summary>
@@ -68,6 +70,21 @@ namespace Amnesia.Handlers {
                 // remove/clean up since protection is inactive
                 player.Buffs.RemoveBuff(Values.BuffBloodmoonLifeProtection);
                 player.Buffs.RemoveBuff(Values.BuffPostBloodmoonLifeProtection);
+            }
+        }
+
+        /// <summary>
+        /// Temporary method to automatically refund any players with the Hardened Memory buff from version 1.0.0.
+        /// </summary>
+        /// <param name="player">EntityPlayer to check buffs for and refund if hardened.</param>
+        private static void RefundHardenedMemory(ClientInfo clientInfo, EntityPlayer player) {
+            if (player.Buffs.HasBuff(Values.BuffHardenedMemory)) {
+                PlayerHelper.GiveItem(clientInfo, player, Values.NameMemoryBoosterItem);
+                MessagingSystem.Whisper("[ff8000]The Amnesia mod was recently updated and some things have changed.[-]", player.entityId);
+                MessagingSystem.Whisper("A [007fff]Memory Booster[-] you previously used [00ff80]is being given back to you[-] since it now serves a slightly different purpose.", player.entityId);
+                MessagingSystem.Whisper("[ff007f]Please check your inventory now to confirm it was received[-]. If you don't see it there, check for a bag on the ground.", player.entityId);
+                MessagingSystem.Whisper("For more info on changes, please take a look at the Amnesia entry in your [ff8000]Journal[-].", player.entityId);
+                player.Buffs.RemoveBuff(Values.BuffHardenedMemory);
             }
         }
     }