+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - ?
 
 - add fragile memory state
+- add retroactive refund for hardened memory
 - remove non-time text from buffs for cleanliness
 - remove references to lives in positive outlook
 - update journal entry to reflect new flow